🚜 Move filterRows function to utils
Export utils function from an index file and separate each funcion.

Show "no records to display" if filtered data is empty.
